+++ backend/free_time/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class FreeTimeConfig(AppConfig):
+    name = 'free_time'
