--- backend/younmeet_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class YounmeetAppConfig(AppConfig):
-    name = 'younmeet_app'
